+++ sealed_annotations/example/mixed.sealed.dart
@@ -52,40 +52,6 @@ abstract class Apple {
     }
   }
 
-  R whenOrDefault<R extends Object?>({
-    R Function(AppleHold hold)? hold,
-    required R orDefault,
-  }) {
-    final apple = this;
-    if (apple is AppleHold) {
-      return hold != null ? hold(apple) : orDefault;
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R? whenOrNull<R extends Object?>({
-    R Function(AppleHold hold)? hold,
-  }) {
-    final apple = this;
-    if (apple is AppleHold) {
-      return hold?.call(apple);
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R whenOrThrow<R extends Object?>({
-    R Function(AppleHold hold)? hold,
-  }) {
-    final apple = this;
-    if (apple is AppleHold && hold != null) {
-      return hold(apple);
-    } else {
-      throw AssertionError();
-    }
-  }
-
   void whenPartial({
     void Function(AppleHold hold)? hold,
   }) {
@@ -159,40 +125,6 @@ abstract class Banana {
     final banana = this;
     if (banana is BananaHold) {
       return hold != null ? hold(banana) : orElse(banana);
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R whenOrDefault<R extends Object?>({
-    R Function(BananaHold hold)? hold,
-    required R orDefault,
-  }) {
-    final banana = this;
-    if (banana is BananaHold) {
-      return hold != null ? hold(banana) : orDefault;
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R? whenOrNull<R extends Object?>({
-    R Function(BananaHold hold)? hold,
-  }) {
-    final banana = this;
-    if (banana is BananaHold) {
-      return hold?.call(banana);
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R whenOrThrow<R extends Object?>({
-    R Function(BananaHold hold)? hold,
-  }) {
-    final banana = this;
-    if (banana is BananaHold && hold != null) {
-      return hold(banana);
     } else {
       throw AssertionError();
     }
@@ -316,58 +248,6 @@ abstract class Coconut {
       return test2 != null ? test2(coconut) : orElse(coconut);
     } else if (coconut is CoconutHold) {
       return hold != null ? hold(coconut) : orElse(coconut);
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R whenOrDefault<R extends Object?>({
-    R Function(CoconutTest1 test1)? test1,
-    R Function(CoconutTest2 test2)? test2,
-    R Function(CoconutHold hold)? hold,
-    required R orDefault,
-  }) {
-    final coconut = this;
-    if (coconut is CoconutTest1) {
-      return test1 != null ? test1(coconut) : orDefault;
-    } else if (coconut is CoconutTest2) {
-      return test2 != null ? test2(coconut) : orDefault;
-    } else if (coconut is CoconutHold) {
-      return hold != null ? hold(coconut) : orDefault;
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R? whenOrNull<R extends Object?>({
-    R Function(CoconutTest1 test1)? test1,
-    R Function(CoconutTest2 test2)? test2,
-    R Function(CoconutHold hold)? hold,
-  }) {
-    final coconut = this;
-    if (coconut is CoconutTest1) {
-      return test1?.call(coconut);
-    } else if (coconut is CoconutTest2) {
-      return test2?.call(coconut);
-    } else if (coconut is CoconutHold) {
-      return hold?.call(coconut);
-    } else {
-      throw AssertionError();
-    }
-  }
-
-  R whenOrThrow<R extends Object?>({
-    R Function(CoconutTest1 test1)? test1,
-    R Function(CoconutTest2 test2)? test2,
-    R Function(CoconutHold hold)? hold,
-  }) {
-    final coconut = this;
-    if (coconut is CoconutTest1 && test1 != null) {
-      return test1(coconut);
-    } else if (coconut is CoconutTest2 && test2 != null) {
-      return test2(coconut);
-    } else if (coconut is CoconutHold && hold != null) {
-      return hold(coconut);
     } else {
       throw AssertionError();
     }