+++ sealed_generators/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Removed support for super_enum API
 - Removed support for non null-safe projects.
+- Removed whenOrThrow, whenOrNull and whenOrDefault
 
 ## 1.4.0